Select/Count/Join/Group SQL Statement
Able to select all the restaurants, and count the number of menus belonging to each one. Therefore, I can see how many menu's 'Recline and Dine' has, as well as how many menus 'The Great Cow' has. Also, through using an alias, I am able to name the columns which contain the menu name, and the restaurant it belongs to.